pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "tumbller-fc-frames-fastapi"
version = "0.1.0"
description = ""
authors = [{"name" = "Anuraj", "email" = "anuraj-rp@users.noreply.github.com"}]
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
  "beautifulsoup4==4.13.4",
  "farcaster==0.7.12",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "jinja2==3.1.6",
  "pillow==11.3.0",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.42",
  "uvicorn==0.35.0",
]

[project.optional-dependencies]
dev = [
  "ruff",
]

prod = [
]
